Update bindings to more recent versions of GSL
Original code was developed with bindings to GSL 1.8 C libraries, and will no longer compile on modern systems.  Here, I've adapted the relevant configure and Makevars files based on the approach used in the current CRAN version of the`gsl` R package (which now provides independent R bindings the the C library, though is not used here.)  These edits allow the package to compile and install, after which, the core functions that reproduce the results in the paper can be run successfully!